+++ cmd/frontend/authz/register.go
