+++ enterprise/cmd/frontend/internal/authz/common.go
@@ -1,4 +1,4 @@
-package shared
+package authz
 
 import (
 	"fmt"
